+++ gDRplots/tests/testthat/test-plot-prism.R
@@ -1006,26 +1006,54 @@ test_that("plot_boxplot_meta works as expected", {
 
 
 test_that("plot_volcano_assoc_panel works as expected", {
-  plt_1 <- plot_volcano_assoc_panel(dt_response = dt_response_met,
+  obj_1 <- plot_volcano_assoc_panel(dt_response = dt_response_met,
                                     dt_depmap = obj_depmap_feat[["dt_depmap"]],
                                     selected_metric = "RV_gDR_x_max",  
                                     selected_feat_meta_col = obj_depmap_feat[["selected_feat_meta_col"]])
+  expect_length(obj_1, 2)
+  expect_is(obj_1, "list")
+  expect_equal(names(obj_1), c("assoc_data", "panel"))
+  plt_1 <- obj_1[["panel"]]
   expect_is(plt_1, "gg")
   expect_true(any(grepl("PANEL", names(ggplot2::ggplot_build(plt_1)[["data"]][[1]]))))
+  tab_1 <- obj_1[["assoc_data"]]
+  expect_is(tab_1, "data.table")
+  expect_equal(names(tab_1), c("feature", "response", "rho", "q_value", "neglog_q_value"))
   
-  plt_2 <- plot_volcano_assoc_panel(dt_response = dt_response_score,
+  obj_2 <- plot_volcano_assoc_panel(dt_response = dt_response_score,
                                     dt_depmap = obj_depmap_feat_2[["dt_depmap"]],
                                     selected_metric = "RV_gDR_bliss_score",  
                                     selected_feat_meta_col = obj_depmap_feat_2[["selected_feat_meta_col"]])
+  expect_is(obj_2, "list")
+  plt_2 <- obj_2[["panel"]]
   expect_is(plt_2, "gg")
   expect_true(any(grepl("PANEL", names(ggplot2::ggplot_build(plt_2)[["data"]][[1]]))))
+  tab_2 <- obj_2[["assoc_data"]]
+  expect_is(tab_2, "data.table")
   
-  plt_3 <- plot_volcano_assoc_panel(dt_response = dt_response_diff,
+  obj_3 <- plot_volcano_assoc_panel(dt_response = dt_response_diff,
                                     dt_depmap = obj_depmap_meta[["dt_depmap"]],
                                     selected_metric = "RV_gDR_x_max_cotrt_diff_0.1_col_fittings",  
                                     selected_feat_meta_col = obj_depmap_meta[["selected_feat_meta_col"]])
+  expect_is(obj_3, "list")
+  plt_3 <- obj_3[["panel"]]
   expect_is(plt_3, "gg")
   expect_true(any(grepl("PANEL", names(ggplot2::ggplot_build(plt_3)[["data"]][[1]]))))
+  tab_3 <- obj_3[["assoc_data"]]
+  expect_is(tab_3, "data.table")
+  
+  # scenario with no data
+  obj_4 <- 
+    plot_volcano_assoc_panel(dt_response = dt_response_diff,
+                             dt_depmap = obj_depmap_meta[["dt_depmap"]],
+                             selected_metric = "RV_gDR_log10_xc50_cotrt_zero_0.001_row_fittings",  
+                             selected_feat_meta_col = obj_depmap_meta[["selected_feat_meta_col"]])
+  expect_is(obj_4, "list")
+  plt_4 <- obj_4[["panel"]]
+  expect_is(plt_4, "gg")
+  expect_true(any(grepl("PANEL", names(ggplot2::ggplot_build(plt_4)[["data"]][[1]]))))
+  tab_4 <- obj_4[["assoc_data"]]
+  expect_null(tab_4, "data.table") # empty data
   
   expect_error(plot_volcano_assoc_panel(dt_response = unlist(dt_response_met),
                                         dt_depmap = obj_depmap_feat[["dt_depmap"]],